Revert "doc/Makefile: drop doc-diff worktree and temporary files on "make clean""
This reverts commit d37da824753fe65fa0714e20b89bf73e7b96933b, which
started to require that we have an executable git available in order
to say "make clean", which gives us a chicken-and-egg problem.

Having to have Git installed, or be in a repository, in order to be
able to run an optional "doc-diff" tool is fine.  Requiring either
in order to run "make clean" is a different story.

Reported by Jonathan Nieder <jrnieder@gmail.com>.