push doc: document the DWYM behavior pushing to unqualified <dst>
Document the DWYM behavior that kicks in when pushing to an
unqualified <dst> reference.

This behavior was added in c5bd078c9d ("Renaming push.", 2005-08-03)
and 3691283c89 ("push: allow unqualified dest refspecs to DWIM",
2008-04-23), and somewhat documented in cc5bbe36f7 ("git-push: Update
description of refspecs and add examples", 2007-06-09), but has never
been fully documented.

The closest we got to having documented it was the description in the
commit message for 3691283c89, which I've borrowed from in writing
this documentation.

Let's also refer to this new documentation from the existing
documentation we had (added in cc5bbe36f7).

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>